finish exercise 2

+++ utils.py
@@ -34,7 +34,7 @@ class CosmicRayDS(InMemoryDataset):
     def process(self):
         f = np.load(self.raw_file_names[0])
         x = torch.tensor(f["data"]).float()
-        y = torch.tensor(f["label"])
+        y = torch.tensor(f["label"]).float()
         n_events, n_points, _ = x.shape
 
         data_list = []
